examples/sdk-python/pyproject.toml
[tool.poetry]
name = "example"
version = "0.1.0"
description = ""
authors = ["Phil Leggetter <phil@leggetter.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
outpost-sdk = "^0.1.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
